move projects uploading function and update so test passes
